requirements.txt
discord.py
Flask
hurry.filesize
uptime
DiscordUtils
psutil